scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio==3.4.3
fake-useragent==1.4.0
scrapy==2.11.0
pytz==2023.3
twython==3.9.1
praw==7.7.1
google-api-python-client==2.108.0
